Move signedNoMoreThan90DaysAgo to domain level
