+++ apps/io-func-sign-user/src/app/use-cases/get-signed-document-content.ts
@@ -1,41 +1,15 @@
 import { pipe } from "fp-ts/lib/function";
-import * as E from "fp-ts/lib/Either";
 import * as TE from "fp-ts/lib/TaskEither";
 import * as A from "fp-ts/lib/Array";
-
-import { differenceInDays } from "date-fns";
 
 import { GetDocumentContent } from "@io-sign/io-sign/document-content";
 import { Document } from "@io-sign/io-sign/document";
 import { EntityNotFoundError } from "@io-sign/io-sign/error";
-import { SignatureRequestSigned } from "@io-sign/io-sign/signature-request";
-import { validate } from "@io-sign/io-sign/validation";
 
-import { SignatureRequest } from "../../signature-request";
-
-export const signedNoMoreThan90DaysAgo = (
-  signatureRequest: SignatureRequest
-): E.Either<Error, SignatureRequestSigned> =>
-  pipe(
-    signatureRequest,
-    validate(
-      SignatureRequestSigned,
-      "The signature request must be in SIGNED status."
-    ),
-    E.chain((signatureRequest) =>
-      pipe(
-        differenceInDays(new Date(), signatureRequest.signedAt),
-        (difference) =>
-          difference < 90
-            ? E.right(signatureRequest)
-            : E.left(
-                new EntityNotFoundError(
-                  "More than 90 days have passed since signing."
-                )
-              )
-      )
-    )
-  );
+import {
+  SignatureRequest,
+  signedNoMoreThan90DaysAgo,
+} from "../../signature-request";
 
 export const makeGetSignedDocumentContent =
   (getDocumentContent: GetDocumentContent) =>
